--- a/Changes.md
+++ b/Changes.md
@@ -1,3 +1,14 @@
+0.59.x.x (relative to 0.59.3.0)
+========
+
+Features
+--------
+
+- Instancer : Added support for generating procedural variation via context variables.
+  - Time offsets for animation.
+  - Seeds for randomisation.
+  - Arbitrary context variables driven by primitive variables.
+
 0.59.3.0 (relative to 0.59.2.0)
 ========