--- a/include/GafferScene/Instancer.h
+++ b/include/GafferScene/Instancer.h
@@ -48,6 +48,43 @@ class GAFFERSCENE_API Instancer : public BranchCreator
 
 	public :
 
+
+		/// Compound plug for representing an image format in a way
+		/// easily edited by users, with individual child plugs for
+		/// each aspect of the format.
+		class ContextVariablePlug : public Gaffer::ValuePlug
+		{
+
+			public :
+
+				GAFFER_PLUG_DECLARE_TYPE( ContextVariablePlug, InstancerContextVariablePlugTypeId, Gaffer::ValuePlug );
+
+				ContextVariablePlug(
+					const std::string &name = defaultName<ContextVariablePlug>(),
+					Direction direction=In,
+					bool defaultEnable = true,
+					unsigned flags = Default
+				);
+
+				~ContextVariablePlug() override;
+
+				/// Accepts no children following construction.
+				bool acceptsChild( const GraphComponent *potentialChild ) const override;
+				Gaffer::PlugPtr createCounterpart( const std::string &name, Direction direction ) const override;
+
+				Gaffer::BoolPlug *enabledPlug();
+				const Gaffer::BoolPlug *enabledPlug() const;
+
+				Gaffer::StringPlug *namePlug();
+				const Gaffer::StringPlug *namePlug() const;
+
+				Gaffer::FloatPlug *quantizePlug();
+				const Gaffer::FloatPlug *quantizePlug() const;
+
+		};
+
+		IE_CORE_DECLAREPTR( ContextVariablePlug );
+
 		Instancer( const std::string &name=defaultName<Instancer>() );
 		~Instancer() override;
 
@@ -99,12 +136,39 @@ class GAFFERSCENE_API Instancer : public BranchCreator
 		Gaffer::BoolPlug *encapsulateInstanceGroupsPlug();
 		const Gaffer::BoolPlug *encapsulateInstanceGroupsPlug() const;
 
+		Gaffer::BoolPlug *seedEnabledPlug();
+		const Gaffer::BoolPlug *seedEnabledPlug() const;
+
+		Gaffer::StringPlug *seedVariablePlug();
+		const Gaffer::StringPlug *seedVariablePlug() const;
+
+		Gaffer::IntPlug *seedsPlug();
+		const Gaffer::IntPlug *seedsPlug() const;
+
+		Gaffer::IntPlug *seedPermutationPlug();
+		const Gaffer::IntPlug *seedPermutationPlug() const;
+
+		Gaffer::BoolPlug *rawSeedPlug();
+		const Gaffer::BoolPlug *rawSeedPlug() const;
+
+		Gaffer::ValuePlug *contextVariablesPlug();
+		const Gaffer::ValuePlug *contextVariablesPlug() const;
+
+		ContextVariablePlug *timeOffsetPlug();
+		const ContextVariablePlug *timeOffsetPlug() const;
+
+		Gaffer::AtomicCompoundDataPlug *variationsPlug();
+		const Gaffer::AtomicCompoundDataPlug *variationsPlug() const;
+
 		void affects( const Gaffer::Plug *input, AffectedPlugsContainer &outputs ) const override;
 
 	protected :
 
 		void hash( const Gaffer::ValuePlug *output, const Gaffer::Context *context, IECore::MurmurHash &h ) const override;
 		void compute( Gaffer::ValuePlug *output, const Gaffer::Context *context ) const override;
+
+		Gaffer::ValuePlug::CachePolicy computeCachePolicy( const Gaffer::ValuePlug *output ) const override;
+		Gaffer::ValuePlug::CachePolicy hashCachePolicy( const Gaffer::ValuePlug *output ) const override;
 
 		bool affectsBranchBound( const Gaffer::Plug *input ) const override;
 		void hashBranchBound( const ScenePath &parentPath, const ScenePath &branchPath, const Gaffer::Context *context, IECore::MurmurHash &h ) const override;
@@ -157,6 +221,12 @@ class GAFFERSCENE_API Instancer : public BranchCreator
 
 		GafferScene::ScenePlug *capsuleScenePlug();
 		const GafferScene::ScenePlug *capsuleScenePlug() const;
+
+		// This plug does heavy lifting when necessary to do an expensive set plug computation
+		// It uses a TaskCollaborate policy to allow threads to cooperate, and is evaluated with
+		// a scenePath in the context to return a PathMatcher for the set contents for one branch
+		Gaffer::PathMatcherDataPlug *setCollaboratePlug();
+		const Gaffer::PathMatcherDataPlug *setCollaboratePlug() const;
 
 		ConstEngineDataPtr engine( const ScenePath &parentPath, const Gaffer::Context *context ) const;
 		void engineHash( const ScenePath &parentPath, const Gaffer::Context *context, IECore::MurmurHash &h ) const;