… python/GafferSceneTest/InstancerTest.py
… python/GafferSceneUI/InstancerUI.py
… src/GafferScene/Instancer.cpp