![instancerContexts3](https://user-images.githubusercontent.com/27969079/104797094-71ac4100-5770-11eb-8e63-3d30ae6ae927.png)

Allow changing the contexts that the prototypes plug is evaluated in - a very powerful feature for instancer setups where the instances differ in more complicated ways that just transforms and attributes, but also a very easy way to make very slow setups.  To avoid users creating incredibly slow setups, we've implemented a display for the current number of unique variations being evaluated, and quantization features to reduce this.

Some of this probably still needs work ( particularly tooltips I haven't edited for brevity yet ), but I've now taken an initial cleanup pass over everything, and I think it's close enough to be worth reviewing.

The number of dummy classes I need to use in UI code really makes me wish there was a way to specify arguments for widgets in the metadata, but maybe that's a topic for a different day.