@danieldresser-ie 2021-01-19
This is still fairly WIP, but I think I've now addressed all feedback except for things were I've replied to the comment with a question, and the tooltips, which I'll tackle first thing tomorrow, but hopefully won't have as many knock-on effects.