@danieldresser-ie 2021-01-20
Getting pretty close to having everything addressed - I've uploaded my most recent commits.  I thought I would have it finished today, but then realized that the unindexed RootPerVertex mode problem that John pointed out is a bit more serious.

The reason that I had thought it was OK was because I had copied the logic from the existing code.  John is correct however that this logic doesn't work.  I've now added in tests for the existing RootPerVertex functionality in the instancer, demonstrating that this functionality has always been broken.  I waited for a bit to post this update to see if a clear solution will occur to me, but all that that has done is confirm that it requires a bit more thought.

It's actually not even just unindexed RootPerVertex mode - the Instancer seems to assume that the user is responsible for uniquefying the input data in all modes.  Setting the prototypeRoots to `IECoreScene.PrimitiveVariable( Vertex, IECore.StringVectorData( [ "/foo", "/foo" ] ), IECore.IntVectorData( [ 0, 0, 1, 1 ] ) )` should probably be OK - the input data is valid, though not optimally uniquefied?  This suggests that Instancer fundamentally should do some uniquefying itself that it doesn't currently do.