@andrewkaufman 2021-01-20
> I've now added in tests for the existing RootPerVertex functionality in the instancer, demonstrating that this functionality has always been broken.

To confirm, I think you meant to say that unindexed primvars in RootPerVertex mode were always broken, but properly uniqueified and indexed primvars are working as expected in that mode, correct?

> It's actually not even just unindexed RootPerVertex mode - the Instancer seems to assume that the user is responsible for uniquefying the input data in all modes.

At least in the past, that was considered a reasonable expectation on the user. Performance was the most important factor, checking the data for uniqueness has a cost, and most DCCs would generate the optimal data in the first place, so our expectation was that cost would be wasteful in practice.

Maybe the philosophy has changed now? Or maybe the risk of non-optimally indexed data outweighs the cost of internal uniquefying in some real production cases? In practice I be surprised to see DCCs generating non-optimal data in the first place. The one caveat to that is trying to generate indexed string data natively in Gaffer doesn't work right now, but that could (and perhaps should) be fixed within OSLObject, rather than handled by Instancer if it costs any performance.