@danieldresser-ie 2021-01-20
> To confirm, I think you meant to say that unindexed primvars in RootPerVertex mode were always broken, but properly uniqueified and indexed primvars are working as expected in that mode, correct?

Correct.

> Or maybe the risk of non-optimally indexed data outweighs the cost of internal uniquefying in some real production cases?

The cost may actually not be very high to make this fully robust.  We just need to test the uniqueness of the part that should already be uniquified anyway - if you have a million rocks using 40 different prototypes, we just need to double check that the 40 different prototypes are actually unique, which shouldn't be expensive, and if they aren't, we can fix it using the code we need anyway for unindexed RootPerVertex.

Unless you're arguing that we should throw if given unindexed roots in RootPerVertex mode, because this data format is less optimal?