@danieldresser-ie 2021-01-21
OK, I think this is basically ready to go.  Simplified the logic by hashing the rootPath for every point - this was a rather good suggestion, since I realized the previous implementation had a crash bug with empty rootPaths ( I first added a test for this ).

I've taken a pass over all the tooltips as well, though I feel like there could still be a few more tweaks needed there.  I don't think we've yet discussed the plug name "seedPossibilities" instead of "numSeeds" either.