@johnhaddon 2021-01-22
Two things I forgot to mention :

1. Failing test : `FAIL: test (GafferSceneUITest.DocumentationTest.DocumentationTest)`. Just a missing docstring.
2. I'll do some performance testing on this before merging on Monday.