@danieldresser-ie 2021-01-22
> I get that it's trying to use seed as a prefix to match the others, but it feels a bit contrived. How about seedTotal? Also, seedEnable should be seedEnabled I think, to match all other enabled plugs.

Hmmm, wish I had caught this question before I talked to you this morning, since I would have discussed it then.  I really don't like seedTotal, since it implies the result of summing, and this isn't that, it's defining the possible range at the beginning of the process.  After a quick brainstorm with Murray, he proposed just calling it "seeds", which actually feels fine to me.  What do you think?

> One other question - should we always put a 0 in each of the variation total fields where they would otherwise be empty? I think that might help explain their function a little better in the default state.

Hmm, the one thing that feels weird to me is that the way the individual sources of variation combine is usually closer to multiplying than adding.  Having timeOffset turned off in a sense contributes 1 possible variation, of None.  It's the same variation contribution you would get if you set the timeOffset variable to a primvar that had all the same values - where you would get 1 variation of timeOffset, and it would mulitply with whatever other variations you are doing.  I certainly don't think it would be right to put a "1" in there when it's turned off, but I think because of this weirdness, I'm more comfortable leaving it blank.

I've squashed the previous commits together, but have left the last round as separate commits.

Remaining to do:
* decide on this last round of questions
* squash
* probably retarget onto 0.59?