@danieldresser-ie 2021-01-27
As per our discussion this morning, I tried to ignore the fact that dealing with sets feels horrible, and just focus on implementing something.  This doesn't feel particularly good, but it does seem to work in my test scene ( I do still need to add this to the test cases ).

Some quick performance numbers, on a test with 8 point clouds, each with 1000000 points, with a long path inside the prototype which is in a set.

Before last commit adding parallelism:
No variations:
Scene generation    4.952s (wall), 6.440s (CPU)
With variations:
Scene generation    28.614s (wall), 30.430s (CPU)

After last commit adding parallelism:
No variations:
Scene generation    2.646s (wall), 22.870s (CPU)
With variations:
Scene generation    9.368s (wall), 78.300s (CPU)

I guess maybe that's not too terrible?  But this is directly evaluating the set - things could definitely get scarier with a downstream network depending on the set, and switching to TaskCollaboration on an internal plug definitely seems appropriate - I guess this would live in BranchCreator?  I would definitely appreciate your thoughts on how to approach this code, I'm not terribly happy with it.